Move ginga viewer into qt sub-directory
