+++ glue/plugins/ginga_viewer/qt/client.py
