+++ glue/plugins/ginga_viewer/qt/tests/test_client.py
