+++ glue/plugins/ginga_viewer/qt/tests/test_viewer_widget.py
