+++ glue/plugins/ginga_viewer/qt/viewer_widget.py
